View Model implemented

+++ app/src/main/java/com/ryalls/team/gofishing/interface/ILaunchDetailView.kt
@@ -1,5 +1,5 @@
 package com.ryalls.team.gofishing.`interface`
 
 interface ILaunchDetailView {
-    fun launchDetailView(dbID : Long)
+    fun launchDetailView(dbID : Int)
 }
